@@ -1,8 +1,8 @@
 // Code generated by gen_go_collector_metrics_set.go; DO NOT EDIT.
-//go:generate go run gen_go_collector_metrics_set.go go1.17
+//go:generate go run gen_go_collector_metrics_set.go go1.18
 
-//go:build go1.17 && !go1.18
-// +build go1.17,!go1.18
+//go:build go1.18 && !go1.19
+// +build go1.18,!go1.19
 
 package prometheus
 
