@@ -38,10 +38,12 @@ func main() {
 		log.Fatal("requires Go version (e.g. go1.17) as an argument")
 	}
 	toolVersion := runtime.Version()
-	if majorVersion := toolVersion[:strings.LastIndexByte(toolVersion, '.')]; majorVersion != os.Args[1] {
-		log.Fatalf("using Go version %q but expected Go version %q", majorVersion, os.Args[1])
+	mtv := majorVersion(toolVersion)
+	mv := majorVersion(os.Args[1])
+	if mtv != mv {
+		log.Fatalf("using Go version %q but expected Go version %q", mtv, mv)
 	}
-	version, err := parseVersion(os.Args[1])
+	version, err := parseVersion(mv)
 	if err != nil {
 		log.Fatalf("parsing Go version: %v", err)
 	}
@@ -91,6 +93,10 @@ func parseVersion(s string) (goVersion, error) {
 	}
 	i, err := strconv.Atoi(s[i+1:])
 	return goVersion(i), err
+}
+
+func majorVersion(v string) string {
+	return v[:strings.LastIndexByte(v, '.')]
 }
 
 func rmCardinality() int {
