@@ -197,13 +197,15 @@ func writeMetrics(w io.Writer, mfs []*dto.MetricFamily, useTags bool, prefix str
 
 	buf := bufio.NewWriter(w)
 	for _, s := range vec {
-		for _, c := range prefix {
-			if _, err := buf.WriteRune(c); err != nil {
+		if prefix != "" {
+			for _, c := range prefix {
+				if _, err := buf.WriteRune(c); err != nil {
+					return err
+				}
+			}
+			if err := buf.WriteByte('.'); err != nil {
 				return err
 			}
-		}
-		if err := buf.WriteByte('.'); err != nil {
-			return err
 		}
 		if err := writeMetric(buf, s.Metric, useTags); err != nil {
 			return err
